Improve code style in Apps page.
