+++ package.json
@@ -28,7 +28,6 @@
     "dayjs": "^1.11.5",
     "framer-motion": "^6",
     "lodash": "^4.0.0",
-    "lodash.debounce": "^4.0.8",
     "next": "12.2.5",
     "next-react-svg": "1.1.3",
     "node-fetch": "^3.2.9",
@@ -42,7 +41,6 @@
     "use-font-face-observer": "^1.2.1"
   },
   "devDependencies": {
-    "@types/lodash.debounce": "^4.0.7",
     "@types/node": "17.0.36",
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.5",
