+++ pages/[network_type]/[network_sub_type]/apps.tsx
@@ -1,8 +1,9 @@
-import Apps from '~/ui/pages/Apps';
-import Page from '~/ui/shared/Page';
-import PageHeader from '~/ui/shared/PageHeader';
 import Head from 'next/head';
 import React from 'react';
+
+import Apps from 'ui/pages/Apps';
+import Page from 'ui/shared/Page';
+import PageHeader from 'ui/shared/PageHeader';
 
 const AppsPage = () => {
   return (
