+++ ui/apps/constants.ts
@@ -1,0 +1,44 @@
+import type { AppCategory } from 'types/client/apps';
+
+export const APP_CATEGORIES: Array<AppCategory> = [
+  {
+    id: 'defi',
+    name: 'DeFi',
+  },
+  {
+    id: 'exchanges',
+    name: 'Exchanges',
+  },
+  {
+    id: 'finance',
+    name: 'Finance',
+  },
+  {
+    id: 'games',
+    name: 'Games',
+  },
+  {
+    id: 'marketplaces',
+    name: 'Marketplaces',
+  },
+  {
+    id: 'nft',
+    name: 'NFT',
+  },
+  {
+    id: 'security',
+    name: 'Security',
+  },
+  {
+    id: 'social',
+    name: 'Social',
+  },
+  {
+    id: 'tools',
+    name: 'Tools',
+  },
+  {
+    id: 'Yield-farming',
+    name: 'yield-farming',
+  },
+];
